Moved SignUp pages to Basic Page folder
