+++ FrontEnd/src/components/SignUp/SignupForm/ResendActivationFormContent.jsx
@@ -6,7 +6,7 @@ import { toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import axios from 'axios';
 import styles from './ResendActivationFormContent.module.css';
-import { EMAIL_PATTERN } from '../../../../../constants/constants';
+import { EMAIL_PATTERN } from '../../../constants/constants';
 
 export function ResendActivationFormContentComponent({ setIsValid }) {
   const navigate = useNavigate();
