+++ FrontEnd/src/components/SignUp/SignupForm/RulesModal.jsx
