+++ FrontEnd/src/components/SignUp/SignupForm/Rules_text.js
