+++ FrontEnd/src/pages/BasicPage/SignUp/ActivateProfilePage.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import { useParams } from 'react-router-dom';
-import AuthorizationPage from '../../Authorization/AuthorizationPage';
+import AuthorizationPage from '../../../components/Authorization/AuthorizationPage';
 import styles from './ActivateProfileModal.module.css';
 
 export function ActivationProfilePage() {
