+++ FrontEnd/src/pages/BasicPage/SignUp/ResendActivationPage.jsx
@@ -1,16 +1,16 @@
 import { Link } from 'react-router-dom';
 import styles from './SignUpPage.module.css';
-import { SignUpFormComponent } from '../components/signup/SignUpForm';
+import { ResendActivationFormComponent } from '../../../components/SignUp/ResendActivationForm';
 
-export function SignUpPage() {
+export function ResendActivationPage() {
   return (
     <div className={styles['sign-up']}>
       <div className={styles['sign-up__body']}>
-        <img className={styles['sign-up-img-left']}
+      <img className={styles['sign-up-img-right']}
           src={`${process.env.REACT_APP_PUBLIC_URL}/img/dots_7x6.png`}
           alt="dots_7x6.png" />
         <div className={styles.container}>
-          <SignUpFormComponent />
+          <ResendActivationFormComponent />
           <div className={styles['sign-in-line']}>
             <div className={styles['sign-in-line__text']}>
               Вже були на нашому сайті?
@@ -20,7 +20,7 @@ export function SignUpPage() {
             </Link>
           </div>
         </div>
-        <img className={styles['sign-up-img-right']}
+        <img className={styles['sign-up-img-left']}
           src={`${process.env.REACT_APP_PUBLIC_URL}/img/dots_7x6.png`}
           alt="dots_7x6.png" />
       </div>
